% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fays_generalized_replication.R
\name{as_fays_gen_rep_design}
\alias{as_fays_gen_rep_design}
\title{Convert a survey design object to a generalized replication replicate design}
\usage{
as_fays_gen_rep_design(
  design,
  variance_estimator = NULL,
  max_replicates = 500,
  psd_option = "warn",
  mse = getOption("survey.replicates.mse"),
  compress = TRUE
)
}
\arguments{
\item{design}{A survey design object created using the 'survey' (or 'srvyr') package,
with class \code{'survey.design'} or \code{'svyimputationList'}.}

\item{variance_estimator}{The name of the variance estimator
whose quadratic form matrix should be created.
See \link[svrep]{variance-estimators} for a
detailed description of each variance estimator.
Options include:
\itemize{
  \item{\strong{"Yates-Grundy"}: }{The Yates-Grundy variance estimator based on
  first-order and second-order inclusion probabilities.}
  \item{\strong{"Horvitz-Thompson"}: }{The Horvitz-Thompson variance estimator based on
  first-order and second-order inclusion probabilities.}
  \item{\strong{"Poisson Horvitz-Thompson"}: }{The Horvitz-Thompson variance estimator
  based on assuming Poisson sampling, with first-order inclusion probabilities
  inferred from the sampling probabilities of the survey design object.}
  \item{\strong{"Stratified Multistage SRS"}: }{The usual stratified multistage variance estimator
  based on estimating the variance of cluster totals within strata at each stage.}
  \item{\strong{"Ultimate Cluster"}: }{The usual variance estimator based on estimating
  the variance of first-stage cluster totals within first-stage strata.}
  \item{\strong{"Deville-1"}: }{A variance estimator for unequal-probability
  sampling without replacement, described in Matei and Tillé (2005)
  as "Deville 1".}
  \item{\strong{"Deville-2"}: }{A variance estimator for unequal-probability
  sampling without replacement, described in Matei and Tillé (2005)
  as "Deville 2".}
  \item{\strong{"SD1"}: }{The non-circular successive-differences variance estimator described by Ash (2014),
  sometimes used for variance estimation for systematic sampling.}
  \item{\strong{"SD2"}: }{The circular successive-differences variance estimator described by Ash (2014).
  This estimator is the basis of the "successive-differences replication" estimator commonly used
  for variance estimation for systematic sampling.}
}}

\item{max_replicates}{The maximum number of replicates to allow (should be as large as possible, given computer memory/storage limitations).
A commonly-recommended default is 500. If the number of replicates needed
for a balanced, fully-efficient estimator is less than \code{max_replicates},
then only the number of replicates needed will be created.
If more replicates are needed than \code{max_replicates}, then the full number of replicates
needed will be created, but only a random subsample will be retained.}

\item{psd_option}{Either \code{"warn"} (the default) or \code{"error"}.
This option specifies what will happen if the target variance estimator
has a quadratic form matrix which is not positive semidefinite. This
can occasionally happen, particularly for two-phase designs. \cr
If \code{psd_option="error"}, then an error message will be displayed. \cr
If \code{psd_option="warn"}, then a warning message will be displayed,
and the quadratic form matrix will be approximated by the most similar
positive semidefinite matrix.
This approximation was suggested by Beaumont and Patak (2012),
who note that this is conservative in the sense of producing
overestimates of variance.
Beaumont and Patak (2012) argue that this overestimation is expected to be
small in magnitude. See \code{\link[svrep]{get_nearest_psd_matrix}}
for details of the approximation.}

\item{mse}{If \code{TRUE}, compute variances from sums of squares around the point estimate from the full-sample weights,
If \code{FALSE}, compute variances from sums of squares around the mean estimate from the replicate weights.}

\item{compress}{This reduces the computer memory required to represent the replicate weights and has no
impact on estimates.}
}
\value{
A replicate design object, with class \code{svyrep.design}, which can be used with the usual functions,
such as \code{svymean()} or \code{svyglm()}.

Use \code{weights(..., type = 'analysis')} to extract the matrix of replicate weights.

Use \code{as_data_frame_with_weights()} to convert the design object to a data frame with columns
for the full-sample and replicate weights.
}
\description{
Converts a survey design object to a replicate design object
with replicate weights formed using the generalized replication method of Fay (1984).
The generalized replication method forms replicate weights
from a textbook variance estimator, provided that the variance estimator
can be represented as a quadratic form whose matrix is positive semidefinite
(this covers a large class of variance estimators).
}
\section{Statistical Details}{

See Fay (1989) for a full description of this replication method,
or see the documentation in \link[svrep]{make_fay_gen_rep_factors} for implementation details.

Let \eqn{v( \hat{T_y})} be the textbook variance estimator for an estimated population total \eqn{\hat{T}_y} of some variable \eqn{y}.
The base weight for case \eqn{i} in our sample is \eqn{w_i}, and we let \eqn{\breve{y}_i} denote the weighted value \eqn{w_iy_i}.
Suppose we can represent our textbook variance estimator as a quadratic form: \eqn{v(\hat{T}_y) = \breve{y}\Sigma\breve{y}^T},
for some \eqn{n \times n} matrix \eqn{\Sigma}.
The only constraint on \eqn{\Sigma} is that, for our sample, it must be symmetric and positive semidefinite.

The replication process creates \eqn{B} sets of replicate weights, where the \eqn{b}-th set of replicate weights is a vector of length \eqn{n} denoted \eqn{\mathbf{a}^{(b)}}, whose \eqn{k}-th value is denoted \eqn{a_k^{(b)}}.
This yields \eqn{B} replicate estimates of the population total, \eqn{\hat{T}_y^{*(b)}=\sum_{k \in s} a_k^{(b)} \breve{y}_k}, for \eqn{b=1, \ldots B}, which can be used to estimate sampling variance.

\deqn{
  v_B\left(\hat{T}_y\right)=\frac{\sum_{b=1}^B\left(\hat{T}_y^{*(b)}-\hat{T}_y\right)^2}{B}
}

This replicate variance estimator can be written as a quadratic form:

  \deqn{
    v_B\left(\hat{T}_y\right) =\mathbf{\breve{y}}^{\prime}\Sigma_B \mathbf{\breve{y}}
  }
  where
  \deqn{
    \boldsymbol{\Sigma}_B = \frac{\sum_{b=1}^B\left(\mathbf{a}^{(b)}-\mathbf{1}_n\right)\left(\mathbf{a}^{(b)}-\mathbf{1}_n\right)^{\prime}}{B}
  }

Note that if the vector of adjustment factors \eqn{\mathbf{a}^{(b)}} has expectation \eqn{\mathbf{1}_n} and variance-covariance matrix \eqn{\boldsymbol{\Sigma}},
then we have the across-replicate expectation \eqn{E_{*}\left( \boldsymbol{\Sigma}_B \right) = \boldsymbol{\Sigma}}. Since the bootstrap process takes the sample values \eqn{\breve{y}} as fixed, the bootstrap expectation of the variance estimator is \eqn{E_{*} \left( \mathbf{\breve{y}}^{\prime}\Sigma_B \mathbf{\breve{y}}\right)= \mathbf{\breve{y}}^{\prime}\Sigma \mathbf{\breve{y}}}.
Thus, we can produce a replication variance estimator with the same expectation as the textbook variance estimator simply by randomly generating \eqn{\mathbf{a}^{(b)}} from a distribution with the following two conditions:
\cr
    \strong{Condition 1}: \eqn{\quad \mathbf{E}_*(\mathbf{a})=\mathbf{1}_n}
\cr
    \strong{Condition 2}: \eqn{\quad \mathbf{E}_*\left(\mathbf{a}-\mathbf{1}_n\right)\left(\mathbf{a}-\mathbf{1}_n\right)^{\prime}=\mathbf{\Sigma}}
\cr \cr
With this function, the matrix \eqn{\boldsymbol{\Sigma}} undergoes a spectral decomposition,
and the eigenvectors and eigenvalues are combined using a Hadamard matrix (see Fay (1984)),
which produces 'balanced' replicates. See
}

\section{Two-Phase Designs}{

For a two-phase design, \code{variance_estimator} should be a list of variance estimators' names,
with two elements, such as \code{list('Ultimate Cluster', 'Poisson Horvitz-Thompson')}.
In two-phase designs, only the following estimators may be used for the second phase:
\itemize{
  \item "Ultimate Cluster"
  \item "Stratified Multistage SRS"
  \item "Poisson Horvitz-Thompson"
}
For statistical details on the handling of two-phase designs,
see the documentation for \link[svrep]{make_twophase_quad_form}.
}

\examples{
library(survey)

## Load an example systematic sample ----
data('library_stsys_sample', package = 'svrep')

## First, ensure data are sorted in same order as was used in sampling
library_stsys_sample <- library_stsys_sample[
  order(library_stsys_sample$SAMPLING_SORT_ORDER),
]

## Create a survey design object
design_obj <- svydesign(
  data = library_stsys_sample,
  strata = ~ SAMPLING_STRATUM,
  ids = ~ 1,
  fpc = ~ STRATUM_POP_SIZE
)

## Convert to generalized replicate design
gen_boot_design_sd2 <- as_gen_boot_design(
  design = design_obj,
  variance_estimator = "SD2",
  replicates = 250, exact_vcov = TRUE
)
gen_rep_design_sd2 <- as_fays_gen_rep_design(
  design = design_obj,
  variance_estimator = "SD2",
  max_replicates = 250,
  mse = TRUE
)

svytotal(x = ~ TOTSTAFF, na.rm = TRUE, design = gen_boot_design_sd2)
svytotal(x = ~ TOTSTAFF, na.rm = TRUE, design = gen_rep_design_sd2)

svyquantile(x = ~ LIBRARIA, quantiles = 0.5, na.rm = TRUE,
            design = gen_boot_design_sd2, interval.type = "quantile")
svyquantile(x = ~ LIBRARIA, quantiles = 0.5, na.rm = TRUE,
            design = gen_rep_design_sd2)
}
\references{
The generalized replication method was first proposed in
Fay (1984). Fay (1989) refined the generalized replication method
to produce "balanced" replicates, in the sense that
each replicate contributes equally to variance estimates.
The advantage of balanced replicates is that one can
still obtain a reasonable variance estimate
by using only a random subset of the replicates.

- Ash, S. (2014). "\emph{Using successive difference replication for estimating variances}."
\strong{Survey Methodology}, Statistics Canada, 40(1), 47–59.
\cr \cr
- Beaumont, Jean-François, and Zdenek Patak. 2012. “On the Generalized Bootstrap for Sample Surveys with Special Attention to Poisson Sampling: Generalized Bootstrap for Sample Surveys.” International Statistical Review 80 (1): 127–48. https://doi.org/10.1111/j.1751-5823.2011.00166.x.
\cr \cr
- Bertail, and Combris. 1997. “Bootstrap Généralisé d’un Sondage.” Annales d’Économie Et de Statistique, no. 46: 49. https://doi.org/10.2307/20076068.
\cr \cr
- Dippo, Cathryn, Robert Fay, and David Morganstein. 1984. “Computing Variances from Complex Samples with Replicate Weights.” In, 489–94. Alexandria, VA: American Statistical Association. http://www.asasrms.org/Proceedings/papers/1984_094.pdf.
\cr \cr
- Fay, Robert. 1984. “Some Properties of Estimates of Variance Based on Replication Methods.” In, 495–500. Alexandria, VA: American Statistical Association. http://www.asasrms.org/Proceedings/papers/1984_095.pdf.
\cr \cr
- Fay, Robert. 1989. “Theory And Application Of Replicate Weighting For Variance Calculations.” In, 495–500. Alexandria, VA: American Statistical Association. http://www.asasrms.org/Proceedings/papers/1989_033.pdf
\cr \cr
- Matei, Alina, and Yves Tillé. (2005).
“\emph{Evaluation of Variance Approximations and Estimators
in Maximum Entropy Sampling with Unequal Probability and Fixed Sample Size.}”
\strong{Journal of Official Statistics}, 21(4):543–70.
}
\seealso{
For greater customization of the method, \code{\link[svrep]{make_quad_form_matrix}} can be used to
represent several common variance estimators as a quadratic form's matrix,
which can then be used as an input to \code{\link[svrep]{make_fay_gen_rep_factors}}.

See \link[svrep]{variance-estimators} for a
description of each variance estimator.
}
