% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successive-difference-replication.R
\name{as_sdr_design}
\alias{as_sdr_design}
\title{Title}
\usage{
as_sdr_design(
  design,
  replicates,
  use_normal_hadamard = FALSE,
  by_stratum = FALSE,
  compress = TRUE,
  mse = getOption("survey.replicates.mse")
)
}
\arguments{
\item{design}{A survey design object created using the 'survey' (or 'srvyr') package,
with class \code{'survey.design'} or \code{'svyimputationList'}.}

\item{replicates}{The number of replicates to form.
Must be a power of 4.}

\item{use_normal_hadamard}{Whether to use a normal Hadamard matrix:
that is, a matrix whose first row and first column only have entries
equal to 1.}

\item{by_stratum}{Whether to form replicates separately for each stratum,
or form replicates for all strata at the same time.}

\item{compress}{Use a compressed representation of the replicate weights matrix.
This reduces the computer memory required to represent the replicate weights and has no
impact on estimates.}

\item{mse}{If \code{TRUE}, compute variances from sums of squares around the point estimate from the full-sample weights,
If \code{FALSE}, compute variances from sums of squares around the mean estimate from the replicate weights.}
}
\value{
A replicate design object, with class \code{svyrep.design}, which can be used with the usual functions,
such as \code{svymean()} or \code{svyglm()}.

Use \code{weights(..., type = 'analysis')} to extract the matrix of replicate weights. \cr
Use \code{as_data_frame_with_weights()} to convert the design object to a data frame with columns
for the full-sample and replicate weights.
}
\description{
Title
}
\examples{

# Load example stratified systematic sample
data('library_stsys_sample', package = 'svrep')

## First, ensure data are sorted in same order as was used in sampling
library_stsys_sample <- library_stsys_sample[
  order(library_stsys_sample$SAMPLING_SORT_ORDER),
]

## Create a survey design object
design_obj <- svydesign(
  data = library_stsys_sample,
  strata = ~ SAMPLING_STRATUM,
  ids = ~ 1,
  fpc = ~ STRATUM_POP_SIZE
)

## Convert to SDR replicate design
sdr_design <- as_sdr_design(
  design = design_obj,
  replicates = 64
)

## Compare to generalized bootstrap
gen_boot_design <- as_gen_boot_design(
  design = design_obj,
  variance_estimator = "SD2",
  replicates = 180,
  exact_vcov = TRUE
)

## Estimate sampling variances
svytotal(x = ~ TOTSTAFF, na.rm = TRUE, design = sdr_design)
svytotal(x = ~ TOTSTAFF, na.rm = TRUE, design = gen_boot_design)
svytotal(x = ~ TOTSTAFF, na.rm = TRUE, design = design_obj)
}
