% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successive-difference-replication.R
\name{assign_hadamard_rows}
\alias{assign_hadamard_rows}
\title{Create a row-assignment matrix for Successive Difference Replication-Method}
\usage{
assign_hadamard_rows(
  n,
  hadamard_order,
  number_of_cycles = ceiling(n/hadamard_order),
  use_first_row = TRUE
)
}
\arguments{
\item{n}{The sample size of the data.}

\item{hadamard_order}{The order of the Hadamard matrix}

\item{number_of_cycles}{The number of cycles to use in the row assignment.
Must be at least as large as \code{n/hadamard_order}.}

\item{use_first_row}{Whether to use the first row of the Hadamard matrix.
The first row of a Hadamard matrix is often all 1's, and so using the first
row to create replicate factors leads to the creation of a replicate
whose weights exactly match the full-sample weights. Thus, using the first row
of the Hadamard matrix may be undesirable for practical purposes,
even if it is valid for the purpose of variance estimation.}
}
\value{
A matrix with \code{n} rows and two columns. Each row
gives the assignment of two rows of a Hadamard matrix to the row of data.
}
\description{
Creates a row assignment matrix:
for each row of a dataset of size \eqn{n}, assigns two rows
of a Hadamard matrix.
}
\examples{
# Assign rows of a 4x4 Hadamard matrix
# to 9 observations, using two cycles
assign_hadamard_rows(
  n = 9,
  hadamard_order = 4,
  number_of_cycles = 2
)

}
\references{
Ash, S. (2014). "\emph{Using successive difference replication for estimating variances}."
\strong{Survey Methodology}, Statistics Canada, 40(1), 47–59.
}
