% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_forms_of_survey_design_objects.R
\name{compress_design}
\alias{compress_design}
\title{Produce a compressed representation of a survey design object}
\usage{
compress_design(design, vars_to_keep = NULL)
}
\arguments{
\item{design}{A survey design object}

\item{vars_to_keep}{(Optional) A character vector
of variables in the design to keep in the compressed design.
By default, none of the variables are retained.}
}
\value{
A list with two elements. The \code{design_subset}
element is a a design object with only the minimal rows
needed to represent the survey design.
The \code{index} element links each row of the original design
to a row of \code{design_subset}, so that the design can be "uncompressed."
}
\description{
Produce a compressed representation of a survey design object
}
\keyword{internal}
