% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized-replication-helpers.R
\name{compute_eigen_decomposition}
\alias{compute_eigen_decomposition}
\title{Compute the eigendecomposition of a matrix}
\usage{
compute_eigen_decomposition(Sigma)
}
\arguments{
\item{Sigma}{A symmetric matrix}
}
\value{
A list object with two named components:
\itemize{
\item \code{values}: A vector of eigenvalues.
\item \code{vectors}: A matrix with eigenvectors corresponding to the eigenvalues.
}
}
\description{
Compute the eigendecomposition of a matrix
}
\details{
Computes the eigendecomposition of a symmetric matrix,
either using the base R function \code{eigen()} or instead using the 'torch'
package if the user has 'torch' installed and
has set \code{options(svrep.torch_device = "cpu")}
or \code{options(svrep.torch_device = "cuda")}.
}
\keyword{internal}
