% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bootstrap_weights.R
\name{draw_antal_tille_resample}
\alias{draw_antal_tille_resample}
\title{Draw a replicate sample using the method
of Antal and Tille (2014)}
\usage{
draw_antal_tille_resample(sel_probs)
}
\arguments{
\item{sel_probs}{A vector of selection probabilities
(i.e., inclusion probabilities) for every unit in the sample.}
}
\value{
A vector the same length as \code{sel_probs},
with elements equal to 0, 1, 2, or 3.
This vector indicate how many times each unit
is resampled.
}
\description{
Draw a replicate sample using the method
described in Section 7 of Antal and Tillé (2014).
}
\references{
Antal, E. and Tillé, Y. (2014). 
"A new resampling method for sampling designs without replacement: The doubled half bootstrap." 
\strong{Computational Statistics}, \emph{29}(5), 1345-1363. https://doi.org/10.1007/s00180-014-0495-0
}
\keyword{internal}
