% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bootstrap_weights.R
\name{draw_doubled_half_sample}
\alias{draw_doubled_half_sample}
\title{Draw a doubled half-sample}
\usage{
draw_doubled_half_sample(n)
}
\arguments{
\item{n}{An integer greater than 1.}
}
\value{
A vector of length \code{n},
with elements equal to 0, 1, 2, or 3.
This vector indicate how many times each unit
is resampled. 

When \code{n} is even,
half the values will be 0 and the other half will be 2.
When \code{n} is odd, the values can be 0, 1, 2, or 3.
}
\description{
Draw a doubled half-sample using the method
described in Antal and Tille (2014)
}
\keyword{internal}
