% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_forms_of_survey_design_objects.R
\name{get_design_quad_form}
\alias{get_design_quad_form}
\title{Determine the quadratic form matrix of a variance estimator for a survey design object}
\usage{
get_design_quad_form(design, variance_estimator)
}
\arguments{
\item{design}{A survey design object created using the 'survey' (or 'srvyr') package,
with class \code{'survey.design'} or \code{'svyimputationList'}. Also accepts two-phase design objects
with class \code{'twophase2'}; see the section below titled "Two-Phase Designs" for
more information about handling of two-phase designs.}

\item{variance_estimator}{The name of the variance estimator
whose quadratic form matrix should be created. \cr
See the section "Variance Estimators" below.
Options include:
\itemize{
  \item{\strong{"Yates-Grundy"}: }{The Yates-Grundy variance estimator based on
  first-order and second-order inclusion probabilities.}
  \item{\strong{"Horvitz-Thompson"}: }{The Horvitz-Thompson variance estimator based on
  first-order and second-order inclusion probabilities.}
  \item{\strong{"Poisson Horvitz-Thompson"}: }{The Horvitz-Thompson variance estimator
  based on assuming Poisson sampling with specified first-order inclusion probabilities.}
  \item{\strong{"Stratified Multistage SRS"}: }{The usual stratified multistage variance estimator
  based on estimating the variance of cluster totals within strata at each stage.}
  \item{\strong{"Ultimate Cluster"}: }{The usual variance estimator based on estimating
  the variance of first-stage cluster totals within first-stage strata.}
  \item{\strong{"SD1"}: }{The non-circular successive-differences variance estimator described by Ash (2014),
  sometimes used for variance estimation for systematic sampling.}
  \item{\strong{"SD2"}: }{The circular successive-differences variance estimator described by Ash (2014).
  This estimator is the basis of the "successive-differences replication" estimator commonly used
  for variance estimation for systematic sampling.}
}}
}
\value{
A matrix representing the quadratic form of a specified variance estimator,
based on extracting information about clustering, stratification,
and selection probabilities from the survey design object.
}
\description{
Determines the quadratic form matrix of a specified variance estimator,
by parsing the information stored in a survey design object created using
the 'survey' package.
}
\section{Variance Estimators}{

The \strong{Horvitz-Thompson} variance estimator:
\deqn{
  v(\hat{Y}) = \sum_{i \in s}\sum_{j \in s} (1 - \frac{\pi_i \pi_j}{\pi_{ij}}) \frac{y_i}{\pi_i} \frac{y_j}{\pi_j}
}
The \strong{Yates-Grundy} variance estimator:
\deqn{
  v(\hat{Y}) = -\frac{1}{2}\sum_{i \in s}\sum_{j \in s} (1 - \frac{\pi_i \pi_j}{\pi_{ij}}) (\frac{y_i}{\pi_i} - \frac{y_j}{\pi_j})^2
}
The \strong{Stratified Multistage SRS} variance estimator is the recursive variance estimator
proposed by Bellhouse (1985) and used in the 'survey' package's function \link[survey]{svyrecvar}.
The estimator can be used for any number of sampling stages. For illustration, we describe its use
for two sampling stages.
\deqn{
  v(\hat{Y}) = \hat{V}_1 + \hat{V}_2
}
where
\deqn{
  \hat{V}_1 = \sum_{h=1}^{H} (1 - \frac{n_h}{N_h})\frac{n_h}{n_h - 1} \sum_{i=1}^{n_h} (y_{hi.} - \bar{y}_{hi.})^2
}
and
\deqn{
  \hat{V}_2 = \sum_{h=1}^{H} \frac{n_h}{N_h} \sum_{i=1}^{n_h}v_{hi}(y_{hi.})
}
where \eqn{n_h} is the number of sampled clusters in stratum \eqn{h},
\eqn{N_h} is the number of population clusters in stratum \eqn{h},
\eqn{y_{hi.}} is the weighted cluster total in cluster \eqn{i} of stratum \eqn{h},
\eqn{\bar{y}_{hi.}} is the mean weighted cluster total of stratum \eqn{h},
(\eqn{\bar{y}_{hi.} = \frac{1}{n_h}\sum_{i=1}^{n_h}y_{hi.}}), and
\eqn{v_{hi}(y_{hi.})} is the estimated sampling variance of \eqn{y_{hi.}}.
\cr \cr
The \strong{Ultimate Cluster} variance estimator is simply the stratified multistage SRS
variance estimator, but ignoring variances from later stages of sampling.
\deqn{
  v(\hat{Y}) = \hat{V}_1
}
This is the variance estimator used in the 'survey' package when the user specifies
\code{option(survey.ultimate.cluster = TRUE)} or uses \code{svyrecvar(..., one.stage = TRUE)}.
When the first-stage sampling fractions are small, analysts often omit the finite population corrections \eqn{(1-\frac{n_h}{N_h})}
when using the ultimate cluster estimator.
\cr \cr
The \strong{SD1} and \strong{SD2} variance estimators are "successive difference"
estimators sometimes used for systematic sampling designs.
Ash (2014) describes each estimator as follows:
\deqn{
  \hat{v}_{S D 1}(\hat{Y}) = \left(1-\frac{n}{N}\right) \frac{n}{2(n-1)} \sum_{k=2}^n\left(\breve{y}_k-\breve{y}_{k-1}\right)^2
}
\deqn{
  \hat{v}_{S D 2}(\hat{Y}) = \left(1-\frac{n}{N}\right) \frac{1}{2}\left[\sum_{k=2}^n\left(\breve{y}_k-\breve{y}_{k-1}\right)^2+\left(\breve{y}_n-\breve{y}_1\right)^2\right]
}
where \eqn{\breve{y}_k = y_k/\pi_k} is the weighted value of unit \eqn{k}
with selection probability \eqn{\pi_k}. The SD1 estimator is recommended by Wolter (1984).
The SD2 estimator is the basis of the successive difference replication estimator commonly
used for systematic sampling designs. See Ash (2014) for details.
\cr \cr
For multistage samples, SD1 and SD2 are applied to the clusters at each stage, separately by stratum.
For later stages of sampling, the variance estimate from a stratum is multiplied by the product
of sampling fractions from earlier stages of sampling. For example, at a third stage of sampling,
the variance estimate from a third-stage stratum is multiplied by \eqn{\frac{n_1}{N_1}\frac{n_2}{N_2}},
which is the product of sampling fractions from the first-stage stratum and second-stage stratum.
}

\section{Two-Phase Designs}{

For a two-phase design, \code{variance_estimator} should be a list of variance estimators,
with two elements, such as \code{list('Ultimate Cluster', 'Poisson Horvitz-Thompson')}.
In two-phase designs, only the following estimators may be used for the second phase:
\itemize{
  \item "Ultimate Cluster"
  \item "Stratified Multistage SRS"
  \item "Poisson Horvitz-Thompson"
}
The two-phase variance estimator has a quadratic form determined as follows.
Suppose the first phase sample has size \eqn{n_a}
and the second phase sample has size \eqn{n_b}.
Let \eqn{\boldsymbol{\Sigma}_a} denote the quadratic form matrix for the variance estimator
used for the full first-phase design, with dimension \eqn{n_a \times n_a}.
Let \eqn{\boldsymbol{\Sigma}_{a^\prime}} denote the matrix of dimension \eqn{n_b \times n_b}
formed by subsetting the rows and columns of \eqn{\boldsymbol{\Sigma}_a} to only include
cases selected in the second-phase sample. We let \eqn{\boldsymbol{\Sigma}_{b}} denote
the matrix of dimension \eqn{n_b \times n_b} representing the Horvitz-Thompson
estimator of variance for the second-phase sample, conditional on the selected
first-phase sample.
The second phase sample has a matrix \eqn{\boldsymbol{D}_b} of weights formed by the inverses of
joint inclusion probabilities, with element \eqn{kl} equal to \eqn{\pi_{bkl}^{-1}},
where \eqn{\pi_{bkl}} is the conditional probability that units \eqn{k} and \eqn{l} are included
in the second-phase sample, given the selected first-phase sample.
The second phase sample also has a diagonal matrix \eqn{\boldsymbol{W}_b}
whose \eqn{k}-th diagonal entry is the second-phase weight \eqn{\pi_{bk}^{-1}},
where \eqn{\pi_{bk}} is the conditional probability that unit \eqn{k}
is included in the second-phase sample, given the selected first-phase sample.
\cr
Then the two-phase variance estimator has a quadratic form matrix \eqn{\boldsymbol{\Sigma}_{ab}} given by:
\deqn{
  \boldsymbol{\Sigma}_{ab} = {W}^{-1}_b(\boldsymbol{\Sigma}_{a^\prime} \circ D_b ){W}^{-1}_b + \boldsymbol{\Sigma}_b
}
The first term estimates the variance contribution from the first phase of sampling,
while the second term estimates the variance contribution from the second phase of sampling.
}

\examples{
# Example 1: Bootstrap based on the successive-difference estimator ----

   data('library_stsys_sample', package = 'svrep')

   ## First, ensure data are sorted in same order as was used in sampling
   library_stsys_sample <- library_stsys_sample[
     order(library_stsys_sample$SAMPLING_SORT_ORDER),
   ]

   ## Create a survey design object
   design_obj <- svydesign(
     data = library_stsys_sample,
     strata = ~ SAMPLING_STRATUM,
     ids = ~ 1,
     fpc = ~ STRATUM_POP_SIZE
   )

   ## Convert to generalized bootstrap replicate design
   gen_boot_design_sd2 <- as_gen_boot_design(
     design = design_obj,
     variance_estimator = "SD2",
     replicates = 2000
   )

# Example 2: Two-phase design (second phase is nonresponse) ----

   ## Create a survey design object
   twophase_design <- twophase(
     data = library_stsys_sample,
     strata = list(~ SAMPLING_STRATUM, NULL),
     id = list(~ 1, ~ 1),
     fpc = list(~ STRATUM_POP_SIZE, NULL),
     subset = ~ I(RESPONSE_STATUS == "Survey Respondent")
   )

  ## Convert to a bootstrap design
  as_gen_boot_design(
    design = twophase_design,
    variance_estimator = list(
      "SD2",
      "Poisson Horvitz-Thompson"
    ),
    replicates = 2000
  )
}
\references{
- Ash, S. (2014). "\emph{Using successive difference replication for estimating variances}."
\strong{Survey Methodology}, Statistics Canada, 40(1), 47–59.
\cr \cr
- Bellhouse, D.R. (1985). "\emph{Computing Methods for Variance Estimation in Complex Surveys}."
\strong{Journal of Official Statistics}, Vol.1, No.3.
\cr \cr
- Särndal, C.-E., Swensson, B., & Wretman, J. (1992). "\emph{Model Assisted Survey Sampling}." Springer New York.
}
