% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_forms_of_survey_design_objects.R
\name{get_rep_design_quad_form}
\alias{get_rep_design_quad_form}
\title{Quadratic Form Matrix of A Replicate Design}
\usage{
get_rep_design_quad_form(rep_design)
}
\arguments{
\item{rep_design}{A replicate survey design object created using the 'survey' (or 'srvyr') package,
with class \code{'svyrep.design'}.}
}
\value{
The quadratic form matrix implied by the replicate weights.
The dimension of the matrix matches the number of observations in the data.
}
\description{
Determines the quadratic form matrix corresponding
to the replicate weights of a survey design object.
}
\examples{
# Create an example survey design object
data(scd)
scd_svy <- svydesign(
  data   = scd, 
  prob   = ~1, 
  id     = ~ambulance, 
  strata = ~ESA,
  nest   = TRUE
)

# Create replicate weights
scd_brr <- scd_svy |> as_bootstrap_design(mse = FALSE)

# Get the quadratic form matrix of the replicate weights
brr_quad_form <- get_rep_design_quad_form(scd_brr)

# Compute the variance-covariance of an estimated total
svytotal(x = ~ arrests, design = scd_brr) |> vcov()

# Check that the quadratic form gives the exact same result
scd$arrests \%*\% brr_quad_form \%*\% scd$arrests

}
