% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-functions.R
\name{get_rep_scale_coefs}
\alias{get_rep_scale_coefs}
\title{Access Replication Scale Coefficients}
\usage{
get_rep_scale_coefs(rep_design, output = "separate")
}
\arguments{
\item{rep_design}{A replicate design object}

\item{output}{Either \code{'separate'} (which is the default)
or \code{'combined'}. If \code{'separate'}, the result is a
list with elements 'scale' and 'rscales', where 'scale' is the 
overall scale coefficient and 'rscales' is a vector of coefficients
for each replicate. If \code{'combined'}, the result is a vector
of coefficients for each replicate, which incorporates the overall
scale coefficient.}
}
\description{
Get the scale coefficents
used for variance estimation in a replicate design object.
}
\details{
For a statistic \eqn{\hat{\theta}},
replication methods estimate the sampling variance
using \eqn{R} replicate estimates, with the estimate
for the \eqn{r}-th replicate denoted \eqn{\hat{\theta}_r}.

The formula for the variance estimate is the following:
\deqn{
  v(\hat{\theta}) = C \sum_{r=1}^{R} c_r (\hat{\theta_r} - \hat{\theta})^2
}

The terms \eqn{C} and \eqn{c_r, r=1,\dots,R} are scale coefficients.
\eqn{C} is the overall coefficient,
and \eqn{c_r, r=1,\dots,R} are replicate-specific coefficients.

When calling \code{get_rep_scale_coefs(output = 'separate')},
the list object includes an element \code{'scale'} corresponding
to \eqn{C} and a vector named \code{'rscales'} corresponding
to \eqn{c_r, r=1,\dots,R}.

When calling \code{get_rep_scale_coefs(output = 'combined')},
the output is a vector with \eqn{R} elements,
where the \eqn{r}-th element is \eqn{C \times c_r}.
}
\examples{
data('api', package = 'survey')

api_design <- svydesign(
  data    = apistrat, 
  id      = ~ 1, 
  strata  = ~ stype,
  weights = ~ pw,
  fpc     = ~ fpc,
  nest    = TRUE
)

jk_design <- api_design |>
  as_random_group_jackknife_design(replicates = 12)

jk_design |>
  get_rep_scale_coefs('separate')

jk_design |>
  get_rep_scale_coefs('combined')

}
