% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-helpers.R
\name{getvars}
\alias{getvars}
\title{Get variables from a database}
\usage{
getvars(
  formula,
  dbconnection,
  tables,
  db.only = TRUE,
  updates = NULL,
  subset = NULL
)
}
\arguments{
\item{formula}{Either a formula or a character vector giving names of variables}

\item{dbconnection}{A database connection}

\item{tables}{Name(s) of table(s) to pull from}

\item{db.only}{Unclear parameter inherited from the 'survey' package}

\item{updates}{Updates to potentially make}

\item{subset}{Optional indices of data to subset when returning result}
}
\value{
A data frame
}
\description{
A database helper function copied from the 'survey' package
}
\keyword{internal}
