% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_forms.R
\name{is_psd_matrix}
\alias{is_psd_matrix}
\title{Check whether a matrix is positive semi-definite, based on checking for symmetric and negative eigenvalues.}
\usage{
is_psd_matrix(X, tolerance = 1e-06)
}
\arguments{
\item{X}{A matrix with no missing or infinite values.}

\item{tolerance}{Computed eigenvalues below \code{-abs(tolerance)}
will be considered negative. Some tolerance is recommended
since eigenvalues are nearly always computed with some floating-point error.}
}
\value{
A logical value. \code{TRUE} if the matrix is deemed positive semi-definite.
Negative otherwise (including if \code{X} is not symmetric).
}
\description{
Check whether a matrix is positive semi-definite, based on checking for symmetric and negative eigenvalues.
}
\keyword{internal}
