% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_forms.R
\name{make_breidt_chauvet_matrix}
\alias{make_breidt_chauvet_matrix}
\title{Create a quadratic form's matrix
for the Breidt-Chauvet variance estimator}
\usage{
make_breidt_chauvet_matrix(probs, aux_vars)
}
\arguments{
\item{probs}{A vector of first-order inclusion probabilities}

\item{aux_vars}{A matrix of auxiliary variables,
with the number of rows matching the number of elements of \code{probs}.}
}
\value{
A symmetric matrix whose dimension matches the length of \code{probs}.
}
\description{
Creates the quadratic form matrix for the
variance estimator of Breidt and Chauvet (2011),
intended to be used in variance estimation for balanced
samples selected using the cube method. See Li, Chen, and Krenzke (2014)
for an explanation of the quadratic form for this estimator
and an example of its use as the basis for a generalized replication
estimator.
}
\references{
- Breidt, F.J. and Chauvet, G. (2011).
"Improved variance estimation for balanced samples drawn via the cube method."
Journal of Statistical Planning and Inference, 141, 411-425.

- Li, J., Chen, S., and Krenzke, T. (2014).
"Replication Variance Estimation for Balanced Sampling: An Application to the PIAAC Study."
Proceedings of the Survey Research Methods Section, 2014: 985–994. Alexandria, VA: American Statistical Association.
http://www.asasrms.org/Proceedings/papers/1984_094.pdf.
}
\keyword{internal}
