% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_bootstrap.R
\name{make_optim_boot_factors}
\alias{make_optim_boot_factors}
\title{Optimization-based Bootstrap Weights}
\usage{
make_optim_boot_factors(
  Sigma,
  num_replicates = Matrix::rankMatrix(Sigma) + 1,
  max_iter = 20000,
  max_loss = 1e-04,
  torch_optimizer = function(params, ...) torch::optim_ignite_adamw(params, lr = 5e-04,
    ...),
  ...,
  .verbose = TRUE
)
}
\arguments{
\item{Sigma}{The matrix of the quadratic form used to represent the variance estimator. Must be positive semidefinite.}

\item{num_replicates}{The number of bootstrap replicates to create.}

\item{max_iter}{The maximum iterations to allow for the optimization algorithm.}

\item{max_loss}{The maximum loss to allow for the optimization algorithm.}

\item{torch_optimizer}{An optimization function from the 'torch' package,
such as \code{optim_ignite_adam}.}

\item{...}{Optional arguments to pass to the optimization function
supplied by \code{torch_optimizer}.}

\item{.verbose}{Whether to periodically show the value of the loss function
every several iterations of the optimization algorithm.}
}
\value{
A matrix with the same number of rows as \code{Sigma},
and the number of columns equal to \code{num_replicates}.
}
\description{
Create bootstrap replicate weights
using an optimization algorithm.
}
\examples{
library(survey)

## Load example data
data('library_stsys_sample', package = 'svrep')

## First, ensure data are sorted in same order as was used in sampling
library_stsys_sample <- library_stsys_sample[
  order(library_stsys_sample$SAMPLING_SORT_ORDER),
]

## Create a survey design object
design_obj <- svydesign(
  data = library_stsys_sample,
  strata = ~ SAMPLING_STRATUM,
  ids = ~ 1,
  fpc = ~ STRATUM_POP_SIZE
)

## Obtain quadratic form
quad_form_matrix <- get_design_quad_form(
  design = design_obj,
  variance_estimator = "SD2"
)

## Make optimization-based bootstrap replicates
replication_factors <- make_optim_boot_factors(
  Sigma = quad_form_matrix,
  max_iter = 15000,
  max_loss = 0.005
)

## Use the replicate factors for variance estimation
optim_boot_design <- svrepdesign(
  data = library_stsys_sample,
  repweights = replication_factors,
  type = "other",
  scale = 1/ncol(replication_factors),
  rscales = rep(1, times = ncol(replication_factors)),
  combined.weights = FALSE,
  weights = ~ I(1/SAMPLING_PROB)
)

optim_boot_design |> svymean(x = ~ LIBRARIA, na.rm = TRUE)
}
