% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_forms.R
\name{make_quad_form_matrix}
\alias{make_quad_form_matrix}
\title{Represent a variance estimator as a quadratic form}
\usage{
make_quad_form_matrix(
  variance_estimator = "Yates-Grundy",
  joint_probs = NULL,
  strata_ids = NULL,
  cluster_ids = NULL,
  strata_samp_sizes = NULL,
  strata_pop_sizes = NULL
)
}
\arguments{
\item{variance_estimator}{The name of the variance estimator
whose quadratic form matrix should be created. Options include:
\itemize{
\item{\strong{"Yates-Grundy"}: }{The Yates-Grundy variance estimator based on
first-order and second-order inclusion probabilities. If this is used,
the argument \code{joint_probs} must also be used.}
\item{\strong{"Horvitz-Thompson"}: }{The Horvitz-Thompson variance estimator based on
first-order and second-order inclusion probabilities. If this is used,
the argument \code{joint_probs} must also be used.}
\item{\strong{"Stratified Multistage SRS"}: }{The usual stratified multistage variance estimator
based on estimating the variance of cluster totals within strata at each stage.
If this option is used, then it is necessary to also use the arguments
\code{strata_ids}, \code{cluster_ids}, \code{strata_samp_sizes}, and \code{strata_pop_sizes}.}
\item{\strong{"Ultimate Cluster"}: }{The usual variance estimator based on estimating
the variance of first-stage cluster totals within first-stage strata.
If this option is used, then it is necessary to also use the arguments
\code{strata_ids}, \code{cluster_ids}, \code{strata_samp_sizes}.
Optionally, to use finite population correction factors, one can also use the argument \code{strata_pop_sizes}.}
\item{\strong{"SD1"}: }{The non-circular successive-differences variance estimator described by Ash (2011),
sometimes used for variance estimation for systematic sampling.}
\item{\strong{"SD2"}: }{The circular successive-differences variance estimator described by Ash (2011).
This estimator is the basis of the "successive-differences replication" estimator commonly used
for variance estimation for systematic sampling.}
}}

\item{joint_probs}{Only used if \code{variance_estimator = "Horvitz-Thompson"}.
This should be a matrix of joint inclusion probabilities.
Element \code{[i,i]} of the matrix is the first-order inclusion probability of unit \code{i},
while element \code{[i,j]} is the joint inclusion probability of units \code{i} and \code{j}.}

\item{strata_ids}{Required if \code{variance_estimator} equals \code{"Stratified Multistage SRS"}
or \code{"Ultimate Cluster"}.
This should be a matrix or data frame of strata IDs. If there are multiple stages of sampling,
then \code{strata_ids} can have multiple columns,
with one column for each level of sampling to be accounted for by the variance estimator.}

\item{cluster_ids}{Required unless \code{variance_estimator} equals \code{"Horvitz-Thompson"} or \code{"Yates-Grundy"}.
This should be a matrix or data frame of cluster IDs. If there are multiple stages of sampling,
then \code{cluster_ids} can have multiple columns,
with one column for each level of sampling to be accounted for by the variance estimator.}

\item{strata_samp_sizes}{Required if \code{variance_estimator} equals \code{"Stratified Multistage SRS"} or
\code{"Ultimate Cluster"}. Can also be used if \code{variance_estimator} equals \code{"SD1"} or \code{"SD2"}.
This should be a matrix or data frame giving cluster sample sizes in each stratum.
If there are multiple stages of sampling,
then \code{strata_samp_sizes} can have multiple columns,
with one column for each level of sampling to be accounted for by the variance estimator.}

\item{strata_pop_sizes}{Required if \code{variance_estimator} equals \code{"Stratified Multistage SRS"},
but can optionally be used if \code{variance_estimator} equals \code{"Ultimate Cluster"}, \code{"SD1"}, or \code{"SD2"}.
If there are multiple stages of sampling,
then \code{strata_pop_sizes} can have multiple columns,
with one column for each level of sampling to be accounted for by the variance estimator.}
}
\value{
The matrix of the quadratic form representing the variance estimator.
}
\description{
Common variance estimators for estimated population totals can be represented as a quadratic form,
\eqn{v(\hat{Y}) = \mathbf{\breve{y}}^{\prime}\mathbf{\Sigma}\mathbf{\breve{y}}},
where \eqn{\breve{y}} is the vector of weighted values, \eqn{y_i/\pi_i, \space i=1,\dots,n}.
Given a choice of variance estimator and information about the sample design,
this function constructs the \eqn{n \times n} matrix of the quadratic form, \eqn{\mathbf{\Sigma}}.
}
\section{Variance Estimators}{

The \strong{Horvitz-Thompson} variance estimator:
\deqn{
  v(\hat{Y}) = \sum_{i \in s}\sum_{j \in s} (1 - \frac{\pi_i \pi_j}{\pi_{ij}}) \frac{y_i}{\pi_i} \frac{y_j}{\pi_j}
}
The \strong{Yates-Grundy} variance estimator:
\deqn{
  v(\hat{Y}) = \sum_{i \in s}\sum_{j \in s} (1 - \frac{\pi_i \pi_j}{\pi_{ij}}) (\frac{y_i}{\pi_i} - \frac{y_j}{\pi_j})^2
}
The \strong{Stratified Multistage SRS} variance estimator is the recursive variance estimator
proposed by Bellhouse (1985) and used in the 'survey' package's function \link[survey]{svyrecvar}.
The estimator can be used for any number of sampling stages. For illustration, we describe its use
for two sampling stages.
\deqn{
  \begin{aligned}
  v(\hat{Y}) &= \hat{V}_1 + \hat{V}_2 \\
  \hat{V}_1 &= \sum_{h=1}^{H} (1 - \frac{n_h}{N_h})\frac{n_h}{n_h - 1} \sum_{i=1}^{n_h} (y_{hi.} - \bar{y}_{hi.})^2 \\
  \hat{V}_2 &= \sum_{h=1}^{H} \frac{n_h}{N_h} \sum_{i=1}^{n_h}v_{hi}(y_{hi.}) \\
  \end{aligned}
}
where:
\deqn{
  \begin{aligned}
  n_h &\textit{ is the number of sampled clusters in stratum }h \\
  N_h &\textit{ is the number of population clusters in stratum }h \\
  y_{hi.} &\textit{ is the weighted cluster total in cluster }i\textit{ of stratum }h \\
  \bar{y}_{hi.} &= \frac{1}{n_h}\sum_{i=1}^{n_h}y_{hi.} \\
  v_{hi}(y_{hi.}) &\textit{ is the estimated sampling variance of }y_{hi.}
  \end{aligned}
}
The \strong{Ultimate Cluster} variance estimator is simply the stratified multistage SRS
variance estimator, but ignoring variances from later stages of sampling.
\deqn{
  v(\hat{Y}) = \hat{V}_1
}
This is the variance estimator used in the 'survey' package when the user specifies
\code{option(survey.ultimate.cluster = TRUE)} or uses \code{svyrecvar(..., one.stage = TRUE)}.
}

\section{Arguments required for each variance estimator}{

Below are the arguments that are required or optional for each variance estimator.\tabular{lrrrrr}{
   variance_estimator \tab joint_probs \tab cluster_ids \tab strata_ids \tab strata_samp_sizes \tab strata_pop_sizes \cr
   Yates-Grundy \tab Required \tab  \tab  \tab  \tab  \cr
   Horvitz-Thompson \tab Required \tab  \tab  \tab  \tab  \cr
   Stratified Multistage SRS \tab  \tab Required \tab Required \tab Required \tab Required \cr
   Ultimate Cluster \tab  \tab Required \tab Required \tab Required \tab Optional \cr
   SD1 \tab  \tab Required \tab Optional \tab Optional \tab Optional \cr
   SD2 \tab  \tab Required \tab Optional \tab Optional \tab Optional \cr
}
}

\examples{
# Example 1: The Horvitz-Thompson Estimator

  data("election", package = "survey")

  ht_quad_form_matrix <- make_quad_form_matrix(variance_estimator = "Horvitz-Thompson",
                                               joint_probs = election_jointprob)
  ##_ Produce variance estimate
  wtd_y <- as.matrix(election_pps$wt * election_pps$Bush)
  t(wtd_y) \%*\% ht_quad_form_matrix \%*\% wtd_y

  ##_ Compare against result from 'survey' package
  svytotal(x = ~ Bush,
           design = svydesign(data=election_pps,
                              variance = "HT",
                              pps = ppsmat(election_jointprob),
                              ids = ~ 1, fpc = ~ p))

# Example 2: Stratified multistage Sample ----

}
