% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_forms.R
\name{make_twophase_quad_form}
\alias{make_twophase_quad_form}
\title{Combine quadratic forms from each phase of a two phase design}
\usage{
make_twophase_quad_form(sigma_1, sigma_2, phase_2_joint_probs)
}
\arguments{
\item{sigma_1}{The quadratic form for the first phase variance estimator,
subsetted to only include cases selected in the phase two sample.}

\item{sigma_2}{The quadratic form for the second phase variance estimator,
conditional on the selection of the first phase sample.}

\item{phase_2_joint_probs}{The matrix of conditional joint
inclusion probabilities for the second phase, given the selected
first phase sample.}
}
\value{
A quadratic form matrix that can be used to estimate
the sampling variance from a two-phase sample design.
}
\description{
This function combines quadratic forms from each phase of a two phase design,
so that the combined variance of the entire two-phase sampling can be estimated.
}
\section{Statistical Details}{

The two-phase variance estimator has a quadratic form matrix \eqn{\boldsymbol{\Sigma}_{ab}} given by:
\deqn{
  \boldsymbol{\Sigma}_{ab} = {W}^{-1}_b(\boldsymbol{\Sigma}_{a^\prime} \circ D_b ){W}^{-1}_b + \boldsymbol{\Sigma}_b
}
The first term estimates the variance contribution from the first phase of sampling,
while the second term estimates the variance contribution from the second phase of sampling. \cr

The full quadratic form of the variance estimator is:
\deqn{
  v(\hat{t_y}) = \breve{\breve{y^{'}}} \boldsymbol{\Sigma}_{ab} \breve{\breve{y}}
}
where the weighted variable \eqn{\breve{\breve{y}}_k = \frac{y_k}{\pi_{ak}\pi_{bk}}},
is formed using the first phase inclusion probability, denoted \eqn{\pi_{ak}}, and
the conditional second phase inclusion probability (given the selected first phase sample),
denoted \eqn{\pi_{bk}}. \cr

The notation for this estimator is as follows: \cr
\itemize{
  \item \eqn{n_a} denotes the first phase sample size.

  \item \eqn{n_b} denotes the second phase sample size.

  \item \eqn{\boldsymbol{\Sigma}_a} denotes the matrix of dimension \eqn{n_a \times n_a}
  representing the quadratic form for the variance estimator
  used for the full first-phase design.

  \item \eqn{\boldsymbol{\Sigma}_{a^\prime}} denotes the matrix of dimension \eqn{n_b \times n_b}
  formed by subsetting the rows and columns of \eqn{\boldsymbol{\Sigma}_a} to only include
  cases selected in the second-phase sample.

  \item \eqn{\boldsymbol{\Sigma}_{b}} denotes
  the matrix of dimension \eqn{n_b \times n_b} representing the Horvitz-Thompson
  estimator of variance for the second-phase sample, conditional on the selected
  first-phase sample.

  \item \eqn{\boldsymbol{D}_b} denotes the \eqn{n_b \times n_b} matrix of weights formed by the inverses of
  the second-phase joint inclusion probabilities, with element \eqn{kl} equal to \eqn{\pi_{bkl}^{-1}},
  where \eqn{\pi_{bkl}} is the conditional probability that units \eqn{k} and \eqn{l} are included
  in the second-phase sample, given the selected first-phase sample.

  \item \eqn{\boldsymbol{W}_b} denotes the diagonal \eqn{n_b \times n_b} matrix
  whose \eqn{k}-th diagonal entry is the second-phase weight \eqn{\pi_{bk}^{-1}},
  where \eqn{\pi_{bk}} is the conditional probability that unit \eqn{k}
  is included in the second-phase sample, given the selected first-phase sample.
}
}

\examples{
\dontrun{

## ---------------------- Example 1 ------------------------##
## First phase is a stratified systematic sample            ##
## Second phase is nonresponse, modeled as Poisson sampling ##
##----------------------------------------------------------##

data('library_stsys_sample', package = 'svrep')

# Determine quadratic form for full first-phase sample variance estimator

  full_phase1_quad_form <- make_quad_form_matrix(
    variance_estimator = "SD2",
    cluster_ids = library_stsys_sample[,'FSCSKEY',drop=FALSE],
    strata_ids = library_stsys_sample[,'SAMPLING_STRATUM',drop=FALSE],
    strata_pop_sizes = library_stsys_sample[,'STRATUM_POP_SIZE',drop=FALSE],
    sort_order = library_stsys_sample$SAMPLING_SORT_ORDER
  )

# Identify cases included in phase two sample
# (in this example, respondents)
  phase2_inclusion <- (
    library_stsys_sample$RESPONSE_STATUS == "Survey Respondent"
  )
  phase2_sample <- library_stsys_sample[phase2_inclusion,]

# Estimate response propensities

  response_propensities <- glm(
    data = library_stsys_sample,
    family = quasibinomial('logit'),
    formula = phase2_inclusion ~ 1,
    weights = 1/library_stsys_sample$SAMPLING_PROB
  ) |>
    predict(type = "response",
            newdata = phase2_sample)

# Estimate conditional joint inclusion probabilities for second phase

  phase2_joint_probs <- outer(response_propensities, response_propensities)
  diag(phase2_joint_probs) <- response_propensities

# Determine quadratic form for variance estimator of second phase
# (Horvitz-Thompson estimator for nonresponse modeled as Poisson sampling)

  phase2_quad_form <- make_quad_form_matrix(
    variance_estimator = "Horvitz-Thompson",
    joint_probs = phase2_joint_probs
  )

# Create combined quadratic form for entire design

 twophase_quad_form <- make_twophase_quad_form(
   sigma_1 = full_phase1_quad_form[phase2_inclusion, phase2_inclusion],
   sigma_2 = phase2_quad_form,
   phase_2_joint_probs = phase2_joint_probs
 )

 combined_weights <- 1/(phase2_sample$SAMPLING_PROB * response_propensities)

# Use for variance estimation

  rep_factors <- make_gen_boot_factors(
    Sigma = twophase_quad_form,
    num_replicates = 500
  )

  library(survey)

  twophase_rep_design <- svrepdesign(
    data = phase2_sample,
    type = 'other',
    repweights = rep_factors,
    weights = combined_weights,
    combined.weights = FALSE,
    scale = attr(rep_factors, 'scale'),
    rscales = attr(rep_factors, 'rscales')
  )

  svymean(x = ~ LIBRARIA, design = twophase_rep_design)
}
}
\references{
See Section 7.5 of Tillé (2020) or Section 9.3 of Särndal, Swensson, and Wretman (1992)
for an overview of variance estimation for two-phase sampling. In the case where
the Horvitz-Thompson variance estimator is used for both phases, the method used in this function
is equivalent to equation (9.3.8) of Särndal, Swensson, and Wretman (1992)
and equation (7.7) of Tillé (2020). However, this function can be used
for any combination of first-phase and second-phase variance estimators,
provided that the joint inclusion probabilities from the second-phase design
are available and are all nonzero.
\cr \cr
- Särndal, C.-E., Swensson, B., & Wretman, J. (1992). "\emph{Model Assisted Survey Sampling}." Springer New York.
\cr \cr
- Tillé, Y. (2020). "\emph{Sampling and estimation from finite populations}." (I. Hekimi, Trans.). Wiley.
}
