% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistribute_weights.R
\name{shift_weight}
\alias{shift_weight}
\title{(Internal function) Shift weight from one set of cases to another}
\usage{
shift_weight(wt_set, is_upweight_case, is_downweight_case)
}
\arguments{
\item{wt_set}{A numeric vector of weights}

\item{is_upweight_case}{A logical vector indicating cases whose weight should be increased}

\item{is_downweight_case}{A logical vector indicating cases whose weight should be decreased}
}
\value{
A numeric vector of adjusted weights, of the same length as \code{wt_set}.
}
\description{
You likely want to use \code{redistribute_weights} instead.
The function \code{shift_weight} is internal to this package and is used only "under-the-hood."
}
\examples{

example_data <- data.frame(weight = c(1.5, 1.25, 1.25, 2, 0.5, 1.5),
                           response_status = c(1, 1, 0, 0, 1, 1),
                           stratum = c('A', 'B', 'A', 'A', 'B', 'B'))

svrep:::shift_weight(wt_set = example_data$weight,
                     is_upweight_case = example_data$response_status == 1,
                     is_downweight_case = example_data$response_status == 0)
}
