% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistribute_weights.R
\name{shift_weight_by_group}
\alias{shift_weight_by_group}
\title{Shift weight from one set of cases to another, potentially separately for each value of a grouping variable.}
\usage{
shift_weight_by_group(wt_set, is_upweight_case, is_downweight_case, by_values)
}
\arguments{
\item{wt_set}{A numeric vector of weights}

\item{is_upweight_case}{A logical vector indicating cases whose weight should be increased}

\item{is_downweight_case}{A logical vector indicating cases whose weight should be decreased}

\item{by_values}{A vector of values for which adjustments should be conducted separately}
}
\value{
A numeric vector of adjusted weights, of the same length as \code{wt_set}.
}
\description{
Shift weight from one set of cases to another, potentially separately for each value of a grouping variable.
}
\examples{

example_data <- data.frame(weight = c(1.5, 1.25, 1.25, 2, 0.5, 1.5),
                           response_status = c(1, 1, 0, 0, 1, 1),
                           stratum = c('A', 'B', 'A', 'A', 'B', 'B'))

shift_weight_by_group(wt_set = example_data$weight,
                      is_upweight_case = example_data$response_status == 1,
                      is_downweight_case = example_data$response_status == 0,
                      by_values = example_data$stratum)
}
