% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_replicates.R
\name{subsample_replicates}
\alias{subsample_replicates}
\title{Retain only a random subset of the replicates in a design}
\usage{
subsample_replicates(design, n)
}
\arguments{
\item{design}{A survey design object, created with either the \code{survey} or \code{srvyr} packages.}

\item{n}{The number of replicates to keep after subsampling}
}
\value{
An updated survey design object, where only a random selection
of the replicates has been retained. The overall 'scale' factor for the design
(accessed with \code{design$scale}) is increased to account for the sampling of replicates.
}
\description{
Randomly subsamples the replicates of survey design object,
to keep only a subset. The scale factor used in estimation is increased
to account for the subsampling.
}
\examples{
library(survey)
set.seed(2023)

# Create an example survey design object

  sample_data <- data.frame(
    STRATUM = c(1,1,1,1,2,2,2,2),
    PSU     = c(1,2,3,4,5,6,7,8)
  )

  survey_design <- svydesign(
    data = sample_data,
    strata = ~ STRATUM,
    ids = ~ PSU,
    weights = ~ 1
  )

  rep_design <- survey_design |>
    as_fays_gen_rep_design(variance_estimator = "Ultimate Cluster")

# Inspect replicates before subsampling

  rep_design |> getElement("repweights")

# Inspect replicates after subsampling

  rep_design |>
    subsample_replicates(n = 4) |>
    getElement("repweights")
}
