% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svrep-package.R
\docType{package}
\name{svrep-package}
\alias{svrep}
\alias{svrep-package}
\title{svrep: Tools for Creating, Updating, and Analyzing Survey Replicate Weights}
\description{
Provides tools for creating and working with survey replicate weights.
}
\keyword{internal}
