% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_forms.R
\name{wls_hat_matrix}
\alias{wls_hat_matrix}
\title{Create the "hat matrix" for weighted least squares regression}
\usage{
wls_hat_matrix(X, w)
}
\arguments{
\item{X}{Matrix of predictor variables, with \code{n} rows}

\item{w}{Vector of weights (should all be nonnegative), of length \code{n}}
}
\value{
An \eqn{n \times n} matrix. This is the "hat matrix" for a WLS regression.
}
\description{
Create the "hat matrix" for weighted least squares regression
}
\keyword{internal}
