% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_jackknife_design.R
\name{as_random_group_jackknife_design}
\alias{as_random_group_jackknife_design}
\title{Convert a survey design object to a random-groups jackknife design}
\usage{
as_random_group_jackknife_design(
  design,
  replicates = 50,
  sampling_fraction = 0,
  group_var_name = ".random_group",
  compress = TRUE,
  mse = getOption("survey.replicates.mse")
)
}
\arguments{
\item{design}{A survey design object created using the 'survey' (or 'srvyr') package,
with class \code{'survey.design'} or \code{'svyimputationList'}.}

\item{replicates}{The number of replicates to create.
Every stratum must have at least this many primary sampling units (PSUs),
or else an error will occur.}

\item{sampling_fraction}{By default, the random groups jackknife
assumes that sampling is with-replacement (i.e., the sampling fraction is ignorable).
However, if PSUs were not stratified or if every stratum had the same sampling fraction,
you can specify a single value of \code{sampling_fraction}, which will
be used to adjust the overall replication scale factor through a finite population correction.}

\item{group_var_name}{(Optional) The name of a new variable created to save
identifiers for which random group each PSU was grouped into
for the purpose of forming replicates.
Specify \code{group_var_name = NULL} to avoid creating the variable in the data.}

\item{compress}{Use a compressed representation of the replicate weights matrix.
This reduces the computer memory required to represent the replicate weights and has no
impact on estimates.}

\item{mse}{If \code{TRUE}, compute variances from sums of squares around the point estimate from the full-sample weights,
If \code{FALSE}, compute variances from sums of squares around the mean estimate from the replicate weights.}
}
\value{
A replicate design object, with class \code{svyrep.design}, which can be used with the usual functions,
such as \code{svymean()} or \code{svyglm()}.

Use \code{weights(..., type = 'analysis')} to extract the matrix of replicate weights. \cr
Use \code{as_data_frame_with_weights()} to convert the design object to a data frame with columns
for the full-sample and replicate weights.
}
\description{
Forms a specified number of jackknife replicates
based on grouping primary sampling units (PSUs)
into random, equal-sized groups. The random-groups jackknife
is only valid for designs where the first-stage sampling fraction
is ignorable or is identical in every stratum.
}
\details{
The data are sorted by first-stage sampling strata,
and then the PSUs in each stratum are randomly arranged.
Groups are then formed by sequentially placing PSUs
into each group.
The first PSU in the dataset is placed into the first group,
the second PSU into the second group, and so on.
Once a PSU has been assigned to the last group,
the process begins again by assigning the next PSU to the first group,
the next PSU to the second group, and so on.
}
\examples{
library(survey)

# Load example data

 data('api', package = 'survey')

 api_strat_design <- svydesign(
   data = apistrat,
   id = ~ 1,
   strata = ~stype,
   weights = ~pw
 )

# Create a random-groups jackknife design

 jk_design <- as_random_group_jackknife_design(
   api_strat_design,
   replicates = 15
 )
 print(jk_design)
}
\references{
See Chapter 4 of Wolter (2007) for an overview of the jackknife,
including the method based on random groups.

- Wolter, Kirk. 2007. "Introduction to Variance Estimation." New York, NY: Springer New York. https://doi.org/10.1007/978-0-387-35099-8.
}
