% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_jackknife_design.R
\name{as_random_group_jackknife_design}
\alias{as_random_group_jackknife_design}
\title{Convert a survey design object to a random-groups jackknife design}
\usage{
as_random_group_jackknife_design(
  design,
  replicates = 50,
  var_strat = NULL,
  var_strat_frac = NULL,
  adj_method = "variance-stratum-psus",
  scale_method = "variance-stratum-psus",
  group_var_name = ".random_group",
  compress = TRUE,
  mse = getOption("survey.replicates.mse")
)
}
\arguments{
\item{design}{A survey design object created using the 'survey' (or 'srvyr') package,
with class \code{'survey.design'} or \code{'svyimputationList'}.}

\item{replicates}{The number of replicates to create
for each variance stratum. The total number of replicates
created is the number of variance strata times \code{replicates}.
Every design stratum must have at least as many primary sampling units (PSUs),
as \code{replicates}.}

\item{var_strat}{Specifies the name of a variable
in the data that defines variance strata to use
for the grouped jackknife. If \code{var_strat = NULL},
then there is effectively only one variance stratum.}

\item{var_strat_frac}{Specifies the sampling fraction
to use for finite population corrections in each
value of \code{var_strat}. Can use either a single number
or a variable in the data corresponding to \code{var_strat}.}

\item{adj_method}{Specifies how to calculate the
replicate weight adjustment factor. These adjustment factors
generally take the following form:
\deqn{
 w_{h i(\tilde{h} g)} =
 \begin{cases} w_{h i} & (h i) \notin S_{\tilde{h}} \\
 a_{h(\tilde{h} g)} w_{h i} & (h i) \in S_{\tilde{h}},(h i) \notin S_{\tilde{h} h g} \\
 0 & (h i) \in S_{\tilde{h} h g}
 \end{cases}
}
The options for \code{adj_method} differ in how
\eqn{a_{h(\tilde{h} g)}} is calculated.
Available options include:
\itemize{
  \item{"variance-stratum-psus" (the default)}{
    The replicate weight adjustment for a unit
    is based on the number of PSUs in its variance stratum.
  }
  \item{"variance-units"}{
    The replicate weight adjustment for a unit
    is based on the number of variance units
    in its variance stratum.
  }
}}

\item{scale_method}{Specifies how to calculate the
scale factor for each replicate.
Available options include:
\itemize{
  \item{"variance-stratum-psus"}{
    The scale factor for a variance unit
    is based on its number of PSUs compared
    to the number of PSUs in its variance stratum.
  }
  \item{"variance-units"}{
    The scale factor for a variance unit is
    based on the number of variance units in
    its variance stratum.
  }
}
When variance units in a variance stratum
have differing numbers of PSUs,
the combination \code{adj_method = "variance-stratum-psus"}
and \code{scale_method = "variance-units"} is
recommended by Valliant, Brick, and Dever (2008),
corresponding to their method \code{"GJ2"}.}

\item{group_var_name}{(Optional) The name of a new variable created to save
identifiers for which random group each PSU was grouped into
for the purpose of forming replicates.
Specify \code{group_var_name = NULL} to avoid creating the variable in the data.}

\item{compress}{Use a compressed representation of the replicate weights matrix.
This reduces the computer memory required to represent the replicate weights and has no
impact on estimates.}

\item{mse}{If \code{TRUE}, compute variances from sums of squares around the point estimate from the full-sample weights,
If \code{FALSE}, compute variances from sums of squares around the mean estimate from the replicate weights.}
}
\value{
A replicate design object, with class \code{svyrep.design}, which can be used with the usual functions,
such as \code{svymean()} or \code{svyglm()}.

Use \code{weights(..., type = 'analysis')} to extract the matrix of replicate weights. \cr
Use \code{as_data_frame_with_weights()} to convert the design object to a data frame with columns
for the full-sample and replicate weights.
}
\description{
Forms a specified number of jackknife replicates
based on grouping primary sampling units (PSUs)
into random, (approximately) equal-sized groups.
}
\details{
Within each value of \code{VAR_STRAT},
the data are sorted by first-stage sampling strata,
and then the PSUs in each stratum are randomly arranged.
Groups are then formed by serially placing PSUs
into each group.
The first PSU in the \code{VAR_STRAT} is placed into the first group,
the second PSU into the second group, and so on.
Once a PSU has been assigned to the last group,
the process begins again by assigning the next PSU to the first group,
the PSU after that to the second group, and so on.

The random-groups jackknife method often referred to as "DAGJK"
corresponds to the options \code{var_strat = NULL},
\code{adj_method = "variance-units"}, and \code{scale_method = "variance-units"}.
The DAGJK method will yield upwardly-biased variance estimates for totals
if the total number of PSUs is not a multiple of the total number of replicates (Valliant, Brick, and Dever 2008).
}
\examples{
library(survey)

# Load example data

 data('api', package = 'survey')

 api_strat_design <- svydesign(
   data = apistrat,
   id = ~ 1,
   strata = ~stype,
   weights = ~pw
 )

# Create a random-groups jackknife design

 jk_design <- as_random_group_jackknife_design(
   api_strat_design,
   replicates = 15
 )
 print(jk_design)
}
\references{
See Section 15.5 of Valliant, Dever, and Kreuter (2018)
for an introduction to the grouped jackknife and
guidelines for creating the random groups.

- Valliant, R., Dever, J., Kreuter, F. (2018).
"Practical Tools for Designing and Weighting Survey Samples, 2nd edition." New York: Springer.

See Valliant, Brick, and Dever (2008)
for statistical details related to the
\code{adj_method} and \code{scale_method} arguments.

- Valliant, Richard, Michael Brick, and Jill Dever. 2008.
"Weight Adjustments for the Grouped Jackknife Variance Estimator."
\emph{Journal of Official Statistics}. 24: 469–88.

See Chapter 4 of Wolter (2007) for additional details of the jackknife,
including the method based on random groups.

- Wolter, Kirk. 2007. "Introduction to Variance Estimation." New York, NY: Springer New York. https://doi.org/10.1007/978-0-387-35099-8.
}
