% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successive-difference-replication.R
\name{as_sdr_design}
\alias{as_sdr_design}
\title{Convert a survey design object to a successive differences replicate design}
\usage{
as_sdr_design(
  design,
  replicates,
  sort_variable = NULL,
  use_normal_hadamard = FALSE,
  compress = TRUE,
  mse = getOption("survey.replicates.mse")
)
}
\arguments{
\item{design}{A survey design object created using the 'survey' (or 'srvyr') package,
with class \code{'survey.design'} or \code{'svyimputationList'}.}

\item{replicates}{The target number of replicates to create.
This will determine the order of the Hadamard matrix to use when
creating replicate factors.
If \code{use_normal_hadamard = TRUE}, then the actual number of replicates will be the
smallest \emph{multiple} of 4 that is greater or equal to the specified value of \code{replicates}.
If \code{use_normal_hadamard = FALSE}, then the actual number of replicates will be the
smallest \emph{power} of 4 that is greater or equal to the specified value of \code{replicates}.}

\item{sort_variable}{To create SDR replicates, the data must
be sorted in the order used in sampling. The name of a sorting variable
can be supplied as a character string. If no variable name is supplied,
then this function assumes that the data are already sorted into the correct order.}

\item{use_normal_hadamard}{Whether to use a normal Hadamard matrix:
that is, a matrix whose first row and first column only have entries
equal to 1.}

\item{compress}{Use a compressed representation of the replicate weights matrix.
This reduces the computer memory required to represent the replicate weights and has no
impact on estimates.}

\item{mse}{If \code{TRUE}, compute variances from sums of squares around the point estimate from the full-sample weights,
If \code{FALSE}, compute variances from sums of squares around the mean estimate from the replicate weights.}
}
\value{
A replicate design object, with class \code{svyrep.design}, which can be used with the usual functions,
such as \code{svymean()} or \code{svyglm()}.

Use \code{weights(..., type = 'analysis')} to extract the matrix of replicate weights. \cr
Use \code{as_data_frame_with_weights()} to convert the design object to a data frame with columns
for the full-sample and replicate weights.
}
\description{
Converts a survey design object to a replicate design object
with replicate weights formed using the successive differences replication (SDR) method.
The SDR method is suitable for designs that use
systematic sampling or finely-stratified sampling designs.
The rows of data in the survey design should be sorted
in the same order that was used for sampling.
}
\examples{

# Load example stratified systematic sample
data('library_stsys_sample', package = 'svrep')

## First, ensure data are sorted in same order as was used in sampling
library_stsys_sample <- library_stsys_sample[
  order(library_stsys_sample$SAMPLING_SORT_ORDER),
]

## Create a survey design object
design_obj <- svydesign(
  data = library_stsys_sample,
  strata = ~ SAMPLING_STRATUM,
  ids = ~ 1,
  fpc = ~ STRATUM_POP_SIZE
)

## Convert to SDR replicate design
sdr_design <- as_sdr_design(
  design = design_obj,
  replicates = 180,
  sort_variable = "SAMPLING_SORT_ORDER",
  use_normal_hadamard = TRUE
)

## Compare to generalized bootstrap
## based on the SD2 estimator that SDR approximates
gen_boot_design <- as_gen_boot_design(
  design = design_obj,
  variance_estimator = "SD2",
  replicates = 180,
  exact_vcov = TRUE
)

## Estimate sampling variances
svytotal(x = ~ TOTSTAFF, na.rm = TRUE, design = sdr_design)
svytotal(x = ~ TOTSTAFF, na.rm = TRUE, design = gen_boot_design)
}
\references{
Ash, S. (2014). "\emph{Using successive difference replication for estimating variances}."
\strong{Survey Methodology}, Statistics Canada, 40(1), 47–59.

Fay, R.E. and Train, G.F. (1995). "\emph{Aspects of Survey and Model-Based Postcensal Estimation of
Income and Poverty Characteristics for States and Counties}." Joint Statistical Meetings,
Proceedings of the Section on Government Statistics, 154-159.
}
