% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistribute_weights.R
\name{redistribute_weights}
\alias{redistribute_weights}
\title{Redistribute weight from one group to another}
\usage{
redistribute_weights(design, reduce_if, increase_if, by)
}
\arguments{
\item{design}{A survey design object, created with either the \code{survey} or \code{srvyr} packages.}

\item{reduce_if}{An expression indicating which cases should have their weights set to zero.
Must evaluate to a logical vector with only values of TRUE or FALSE.}

\item{increase_if}{An expression indicating which cases should have their weights increased.
Must evaluate to a logical vector with only values of TRUE or FALSE.}

\item{by}{(Optional) A character vector with the names of variables used to group the redistribution of weights.
For example, if the data include variables named \code{"stratum"} and \code{"wt_class"}, one could specify \code{by = c("stratum", "wt_class")}.}
}
\value{
The survey design object, but with updated full-sample weights and updated replicate weights.
}
\description{
Redistributes weight from one group to another: for example, from non-respondents to respondents.
Redistribution is conducted for the full-sample weights as well as each set of replicate weights.
This can be done separately for each combination of a set of grouping variables, for example to implement a nonresponse weighting class adjustment.
}
\examples{

# Load example data
suppressPackageStartupMessages(library(survey))
data(api)

dclus1 <- svydesign(id=~dnum, weights=~pw, data=apiclus1, fpc=~fpc)
dclus1$variables$response_status <- sample(x = c("Respondent", "Nonrespondent", "Ineligible", "Unknown eligibility"),
                                           size = nrow(dclus1),
                                           replace = TRUE)
rep_design <- as.svrepdesign(dclus1)

# Adjust weights for cases with unknown eligibility
ue_adjusted_design <- redistribute_weights(design = rep_design,
                                           reduce_if = response_status \%in\% c("Unknown eligibility"),
                                           increase_if = !response_status \%in\% c("Unknown eligibility"),
                                           by = c("stype", "cname"))

# Adjust weights for nonresponse
nr_adjusted_design <- redistribute_weights(design = ue_adjusted_design,
                                           reduce_if = response_status \%in\% c("Nonrespondent"),
                                           increase_if = response_status == "Respondent",
                                           by = c("stype", "cname"))
}
