% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svrep-package.R
\docType{package}
\name{svrep-package}
\alias{svrep}
\alias{svrep-package}
\title{svrep: Tools for Creating, Updating, and Analyzing Survey Replicate Weights}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools for creating and working with survey replicate weights, extending functionality of the 'survey' package from Lumley (2004) \doi{10.18637/jss.v009.i08}. Implements bootstrap methods for complex surveys, including the generalized survey bootstrap as described by Beaumont and Patak (2012) \doi{10.1111/j.1751-5823.2011.00166.x}. Methods are provided for applying nonresponse adjustments to both full-sample and replicate weights as described by Rust and Rao (1996) \doi{10.1177/096228029600500305}. Implements methods for sample-based calibration described by Opsomer and Erciulescu (2021) \url{https://www150.statcan.gc.ca/n1/pub/12-001-x/2021002/article/00006-eng.htm}. Diagnostic functions are included to compare weights and weighted estimates from different sets of replicate weights.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bschneidr.github.io/svrep/}
  \item \url{https://github.com/bschneidr/svrep}
  \item Report bugs at \url{https://github.com/bschneidr/svrep/issues}
}

}
\author{
\strong{Maintainer}: Ben Schneider \email{benjamin.julius.schneider@gmail.com} (\href{https://orcid.org/0000-0002-0406-8470}{ORCID})

}
\keyword{internal}
